requirements.txt
munch
requests
dataclasses
dataclasses_json
orjson
numpy
pyyaml
matplotlib
networkx
plotly
kaleido # for plotly